Initial stab at fixing race condition
I was seeing this race condition happen a lot, and essentially
what's happening is this:

   ------> tar -> upload
   ------> tell build API that URL is ready
   !! URL isn't ready sometimes since we did the 2 ops concurrently

The change was to only tell build API about the URL when the
upload is done.

I'm not entirely happy about the use of a callback in
`uploadCwdToSource` but trying to fit in the promise to that
function signature was even more trouble than its worth.